Refactor MealManagementTab into smaller components and fix Fade scrollTop issue
